7/2

+++ Math/GCD.cc
@@ -1,4 +1,4 @@
 //gcd(x, y) -> xとyの最大公約数
-int gcd(int x, int y){
+ll gcd(ll x, ll y){
     return y ? gcd(y, x % y) : x;
 }
